fix(widget): add design mode details and use static resources provided by HunterPie
