+++ ItemBoxTracker/src/Plugin/Views/InventoryView.xaml
@@ -1,37 +1,24 @@
-﻿<src:Widget
+﻿<UserControl
     x:Class="MHWItemBoxTracker.Views.InventoryView"
-    xmlns:src="clr-namespace:HunterPie.GUI;assembly=HunterPie.UI"
     xmlns:Custom="clr-namespace:HunterPie.GUIControls.Custom_Controls;assembly=HunterPie.UI"
     xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
     xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
     xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
     xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
     mc:Ignorable="d"
-    ResizeMode="NoResize"
-    WindowStyle="None"
-    AllowsTransparency="True"
-    ShowInTaskbar="False"
-    Topmost="True"
-    SizeToContent="Height"
-    Width="350"
-    WidgetActive="True">
-  <Window.Resources>
+    Width="350">
+  <UserControl.Resources>
     <ResourceDictionary>
       <ResourceDictionary.MergedDictionaries>
         <ResourceDictionary Source="../Resource/InventoryItemDataTemplate.xaml" />
         <ResourceDictionary Source="../Resource/ItemDataTemplate.xaml" />
       </ResourceDictionary.MergedDictionaries>
     </ResourceDictionary>
-  </Window.Resources>
-  <Window.Background>
-    <SolidColorBrush
-        Color="#00000000"
-        Opacity="0" />
-  </Window.Background>
+  </UserControl.Resources>
   <ItemsControl
       BorderThickness="0"
       ItemsSource="{Binding Items}"
       HorizontalContentAlignment="Stretch"
       Background="{x:Null}"
       IsEnabled="False" />
-</src:Widget>
+</UserControl>
